remove old classes and interfaces
