+++ multiversx_sdk/__init__.py
@@ -5,19 +5,15 @@ from multiversx_sdk.core.account import AccountNonceHolder
 from multiversx_sdk.core.address import (Address, AddressComputer,
                                          AddressFactory)
 from multiversx_sdk.core.code_metadata import CodeMetadata
-from multiversx_sdk.core.contract_query import ContractQuery
-from multiversx_sdk.core.contract_query_builder import ContractQueryBuilder
 from multiversx_sdk.core.message import Message, MessageComputer
 from multiversx_sdk.core.smart_contract_queries_controller import \
     SmartContractQueriesController
 from multiversx_sdk.core.smart_contract_query import (
     SmartContractQuery, SmartContractQueryResponse)
-from multiversx_sdk.core.token_payment import TokenPayment
 from multiversx_sdk.core.tokens import (Token, TokenComputer,
                                         TokenIdentifierParts, TokenTransfer)
 from multiversx_sdk.core.transaction import Transaction
 from multiversx_sdk.core.transaction_computer import TransactionComputer
-from multiversx_sdk.core.transaction_payload import TransactionPayload
 from multiversx_sdk.core.transactions_factories.account_transactions_factory import \
     AccountTransactionsFactory
 from multiversx_sdk.core.transactions_factories.delegation_transactions_factory import \
@@ -67,9 +63,8 @@ from multiversx_sdk.wallet.validator_verifier import ValidatorVerifier
 
 __all__ = [
     "AccountNonceHolder", "Address", "AddressFactory", "AddressComputer",
-    "Transaction", "TransactionPayload", "TransactionComputer",
-    "Message", "MessageComputer", "CodeMetadata", "TokenPayment",
-    "ContractQuery", "ContractQueryBuilder",
+    "Transaction", "TransactionComputer",
+    "Message", "MessageComputer", "CodeMetadata",
     "Token", "TokenComputer", "TokenTransfer", "TokenIdentifierParts",
     "TokenManagementTransactionsOutcomeParser", "SmartContractResult",
     "TransactionEvent", "TransactionLogs", "TransactionOutcome",
