+++ multiversx_sdk/adapters/query_runner_adapter.py
@@ -1,7 +1,6 @@
 from typing import List, Optional, Protocol, Sequence
 
 from multiversx_sdk.core.address import Address
-from multiversx_sdk.core.contract_query import ContractQuery
 from multiversx_sdk.core.interfaces import IAddress
 from multiversx_sdk.core.smart_contract_query import (
     SmartContractQuery, SmartContractQueryResponse)
@@ -60,3 +59,32 @@ class QueryRunnerAdapter:
             return_message=adapted_query_response.return_message,
             return_data_parts=adapted_query_response.get_return_data_parts()
         )
+
+
+class ContractQuery:
+    def __init__(self,
+                 contract: IAddress,
+                 function: str,
+                 encoded_arguments: List[str] = [],
+                 value: Optional[int] = None,
+                 caller: Optional[IAddress] = None) -> None:
+        self.contract: IAddress = contract
+        self.function: str = function
+        self.encoded_arguments = encoded_arguments
+        self.value: int = value or 0
+        self.caller: Optional[IAddress] = caller
+
+    def get_contract(self) -> IAddress:
+        return self.contract
+
+    def get_function(self) -> str:
+        return self.function
+
+    def get_encoded_arguments(self) -> Sequence[str]:
+        return self.encoded_arguments
+
+    def get_caller(self) -> Optional[IAddress]:
+        return self.caller
+
+    def get_value(self) -> int:
+        return self.value
